requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
executing==2.2.0
faiss-cpu==1.12.0
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.19.1
Flask==3.1.2
fqdn==1.5.1
fsspec==2025.7.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
narwhals==2.6.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.3
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
openai==1.97.1
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.32.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pymongo==4.14.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
Send2Trash==1.8.3
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
youtube-transcript-api==1.2.3
yt-dlp==2025.8.11
pandas==2.2.3
serpapi
python-dotenv
trafilatura
PyPDF2
python-docx
firebase-admin